refactor: move Migration Step to the migration file
